Remove old Datatable component + cleanup of new MUI datatable component
