--- src/views/Visualizer/View/DataTable/Header/index.js
@@ -1,9 +1,0 @@
-import Header from './Header';
-import { connectView } from '../../context';
-
-export default connectView(({ setLayerState, layersTreeState }) => ({
-  setLayerState,
-  displayedLayer: Array
-    .from(layersTreeState)
-    .find(([, { table }]) => table),
-}))(Header);
