+++ src/views/Visualizer/View/DataTable/index.js
@@ -1,13 +1,14 @@
 import { connectView } from '../context';
-
 import DataTable from './DataTable';
 
-export default connectView(({ layersTreeState, query, map, visibleBoundingBox }) => ({
-  query,
-  map,
-  visibleBoundingBox,
-  displayedLayer: Array
-    .from(layersTreeState)
-    .filter(([, { table }]) => table)
-    .map(([layer, state]) => ({ ...layer, state }))[0],
-}))(DataTable);
+export default connectView(
+  ({ layersTreeState, query, map, visibleBoundingBox, setLayerState }) => ({
+    query,
+    map,
+    visibleBoundingBox,
+    setLayerState,
+    displayedLayer: Array.from(layersTreeState)
+      .filter(([, { table }]) => table)
+      .map(([layer, state]) => ({ ...layer, state }))[0],
+  }),
+)(DataTable);
