--- src/views/Visualizer/View/DataTableTanstack/styles.scss
@@ -1,17 +1,0 @@
-.data-table-tanstack {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  overflow: hidden;
-
-  .MuiTableContainer-root {
-    flex: 1;
-    min-height: 0;
-    overflow: auto;
-  }
-
-  .MuiTablePagination-root {
-    flex-shrink: 0;
-    border-top: 1px solid rgba(224, 224, 224, 1);
-  }
-}
