+++ src/views/Visualizer/View/Table/styles.scss
